==> STEP 1. Preprocessing the data.ipynb <==
# Raw news texts - preprocessing #

import os
import pandas as pd

# make a list of the raw news text files collected into a directory
# ensuring that Mac's autogenertaed, hidden '.DS_Store' file is not inlcuded in this list
news_files = [f for f in os.listdir(
    "/Users/veronikajuhasz/Documents/University/Cognitive_science_AU/Semester_5/Thesis/analysis/bachelor_project_final/data/NOW_raw")
         if not f.startswith('.') | f.startswith('N')]

news_files

# checking current directory
%pwd

# changing directory to the news_file list's directory
%cd "/Users/veronikajuhasz/Documents/University/Cognitive_science_AU/Semester_5/Thesis/analysis/bachelor_project_final/data/NOW_raw"

# importing multiple raw news txt files into 1 pandas df
news_df =  pd.concat([pd.read_csv(item, sep = '\t', names=["text"]) for item in news_files], axis=0, ignore_index=True)

print(len(news_df))

# extracting the ID number from the beginning of each entry


# putting the ID number into a new column called 'ID'
news_df['ID'] = news_df["text"].str.split(' ').str[0]

# dropping the @@ from the beginning of each ID, with the help of regex
# guide: https://stackoverflow.com/questions/44117326/how-can-i-remove-all-non-numeric-characters-from-all-the-values-in-a-particular)
news_df['ID'] = news_df['ID'].str.extract('(\d+)', expand = False)

# dropping first row without a proper ID - this was not needed after all in this specific file
news_df = news_df[pd.to_numeric(news_df['ID'], errors='coerce').notnull()]

#checking whether ID is numeric, if not, making it into 'int'
print(news_df.dtypes)
news_df["ID"] = news_df["ID"].astype(int)
print(news_df.dtypes)


# checking how the final outcome of these modifications to the news_df dataframe look like
news_df

### Text cleaning ###

# checking one random newstext, what kind of weird systematic symbols are in it?
news_df.iloc[7]['text']

# text cleaning of one specific news text

import re
text = news_df.iloc[7]['text']

# regexing into a clean text:
# 1. remove some special characters
# 2. remove <p>
# 3. remove <h>
# 4. remove 8 digit numbers (ID)
# 5. remove unneccessary space in front of .:,?!'

text_cleaned = re.sub(r"[@*#()&_^]|<p[^>]*>|<h[^>]*>|\d{8}|\s+(?=[.:,?!'])|\s+(?=[.:,?!'])|\\\\", "", text)

# 6. replace many spaces next to each other, replace with 1 space
text_cleaned = re.sub(r"\s\s+", " ", text_cleaned)


text_cleaned

# making succesful text cleaning into a function

def text_cleaner(text):
    text_cleaned = re.sub(r"[@*#()&_^]|<p[^>]*>|<h[^>]*>|\d{8}|\s+(?=[.:,?!'])|\s+(?=[.:,?!'])", "", text)
    text_cleaned = re.sub(r"\s\s+", " ", text_cleaned)   
    return text_cleaned


# testing text cleaner function

selected_text = news_df.iloc[7]['text']
#selected_text

example_cleaned = text_cleaner(selected_text)
print(example_cleaned)

# applying text cleaning function to news dataset

news_df["text"] = news_df.apply(lambda row:
                                text_cleaner(row["text"]), axis = 1)

news_df

 ## adding ".." to the end of each text, so that SVO extractor works later
news_df['text'] = news_df['text'].astype(str) + '..'

# looking at 1 news text to see final look

news_df.iloc[7]['text']

# Source files - preprocessing #

# make a list of the source files collected into a directory
# ensuring that Mac's autogenertaed, hidden '.DS_Store' file is not inlcuded in this list
source_files = [f for f in os.listdir(
    "/Users/veronikajuhasz/Documents/University/Cognitive_science_AU/Semester_5/Thesis/analysis/bachelor_project_final/data/NOW_raw/NOW_source_files")
         if not f.startswith('.')]

source_files

# changing directory to the source_file list's directory
%cd "/Users/veronikajuhasz/Documents/University/Cognitive_science_AU/Semester_5/Thesis/analysis/bachelor_project_final/data/NOW_raw/NOW_source_files"

# importing multiple source txt files into 1 pandas df
source_df =  pd.concat([pd.read_csv(item, sep='\t', names=["ID", "weird_id", "Date", "Country", "News_source_name", "link", "title"]) for item in source_files], axis=0, ignore_index=True)

print(len(source_df))

# filtering news sources for US based news
US_source_df = source_df[source_df["Country"] == "US"]

US_source_df

# finding out how many and what kind of duplicates there are in the source df

print(US_source_df.ID.duplicated().sum())
print(US_source_df.weird_id.duplicated().sum())
print(US_source_df.link.duplicated().sum())
print(US_source_df.title.duplicated().sum())


# Merging news dataset with sources dataset #

# merge the news dataframe with the sources dataframe, then dropping nan-s to filter for US news

news_sources_merged = news_df.merge(US_source_df, on='ID', how='left')

print(news_sources_merged.dtypes)

news_sources_merged

# checking again duplicate news
# we check now as earlier the text column still inlcided unique ID, so the news texts were all unique
# removing news texts that are duplicate in their text

news_sources_merged.text.duplicated().sum()

news_sources_merged = news_sources_merged.drop_duplicates("text").reset_index(drop=True)

news_sources_merged

## Filtering for BLM news ## 

# filtering for news that include BLM relevant terms

term_list = ["Breonna Taylor", "George Floyd", "George Zimmerman", "Black Lives Matter", "BLM", "#BlackLivesMatter", "National Guard", "Derek Chauvin", "Blue Lives Matter", "#BlueLivesMatter", "White Lives Matter", "#WhiteLivesMatter", "All Lives Matter", "#AllLivesMatter", "police brutality"]


term_list_df = news_sources_merged[news_sources_merged.apply(lambda r: any([kw in r[0] for kw in term_list]), axis=1)]
term_list_df = term_list_df.reset_index(level = None, drop = False)


term_list_df

# exporting to csv file on my computer

term_list_df.to_csv("/Users/veronikajuhasz/Documents/University/Cognitive_science_AU/Semester_5/Thesis/analysis/bachelor_project_final/data/NOW_preprocessed/BLM_filtered_preprocessed_news_sources_df.csv", index = False)
