==> STEP 2. SV, VO extraction.ipynb <==
import pandas as pd

term_list_df = pd.read_csv("data/NOW_preprocessed/BLM_filtered_preprocessed_news_sources_df.csv")
print(len(term_list_df))

content = term_list_df["text"][0]

term_list_df

## SV extraction ##

### applying SV extractor (self made from prev. trial 18 versions)
# it is based on the textpipliner package: https://github.com/krzysiekfonal/textpipeliner
# decoding of pipes: https://github.com/krzysiekfonal/grammaregex
# decoding of POS: https://www.guru99.com/pos-tagging-chunking-nltk.html

# adding additional named entity filters to pipe structure
# list of filters: https://stackoverflow.com/questions/59319207/ner-entity-recognition-country-filter

from textpipeliner import PipelineEngine, Context
from textpipeliner.pipes import *
import spacy

nlp = spacy.load('en_core_web_lg')
doc = nlp(content)

def SV_extractor(content):
    doc = nlp(content)
    pipes_structure = [
        AnyPipe([
            SequencePipe([
                FindTokensPipe("VERB/nsubj/*"),
                AggregatePipe([
                    NamedEntityFilterPipe("PERSON"),
                    NamedEntityFilterPipe("FAC"),
                    NamedEntityFilterPipe("LOC"),
                    NamedEntityFilterPipe("GPE"),
                    NamedEntityFilterPipe("PRODUCT"),
                    NamedEntityFilterPipe("LAW"),
                    NamedEntityFilterPipe("LANGUAGE"),
                    NamedEntityFilterPipe("DATE"),
                    NamedEntityFilterPipe("TIME"),
                    NamedEntityFilterPipe("PERCENT"),
                    NamedEntityFilterPipe("MONEY"),
                    NamedEntityFilterPipe("QUANTITY"),
                    NamedEntityFilterPipe("ORDINAL"),
                    NamedEntityFilterPipe("CARDINAL"),
                    NamedEntityFilterPipe("ORG"),
                    NamedEntityFilterPipe("WORK_OF_ART"),
                    NamedEntityFilterPipe("EVENT")
                    ]),
                NamedEntityExtractorPipe()]),
            SequencePipe([
                FindTokensPipe("VERB/agent/*/pobj/*"),
                AggregatePipe([
                    NamedEntityFilterPipe("PERSON"),
                    NamedEntityFilterPipe("FAC"),
                    NamedEntityFilterPipe("LOC"),
                    NamedEntityFilterPipe("GPE"),
                    NamedEntityFilterPipe("PRODUCT"),
                    NamedEntityFilterPipe("LAW"),
                    NamedEntityFilterPipe("LANGUAGE"),
                    NamedEntityFilterPipe("DATE"),
                    NamedEntityFilterPipe("TIME"),
                    NamedEntityFilterPipe("PERCENT"),
                    NamedEntityFilterPipe("MONEY"),
                    NamedEntityFilterPipe("QUANTITY"),
                    NamedEntityFilterPipe("ORDINAL"),
                    NamedEntityFilterPipe("CARDINAL"),
                    NamedEntityFilterPipe("ORG"),
                    NamedEntityFilterPipe("WORK_OF_ART"),
                    NamedEntityFilterPipe("EVENT")
                    ]),
                NamedEntityExtractorPipe()]),
            SequencePipe([
                FindTokensPipe("VERB/nsubj/*")]),
            SequencePipe([
                FindTokensPipe("VERB/agent/*/pobj/*")])]),
        FindTokensPipe("VERB")
    ]
    engine = PipelineEngine(pipes_structure, Context(doc), [0, 1])
    SVs = engine.process()
    return SVs

# applying the SV_extractor function on each row of my pandas dataframe

term_list_df["SVs"] = term_list_df.apply(lambda row: SV_extractor(row["text"]), axis = 1)

SV_df = term_list_df[["ID", "Date", "News_source_name", "SVs"]] 

# within the dataframe, separating the list of SV pairs into individual SV pairs per row

SV_df_exploded = SV_df.explode("SVs")

SV_df_exploded[["S", "V"]] = pd.DataFrame(SV_df_exploded.SVs.tolist(), index = SV_df_exploded.index)

SV_df_exploded

# saving df with VO triples into a csv file

SV_df_exploded.to_csv("data/NOW_preprocessed/total_SV_df.csv", index = False)

## VO extraction ## 

### applying VO extractor (self made from prev. trial 18 versions)
# it is based on the textpipliner package: https://github.com/krzysiekfonal/textpipeliner
# decoding of pipes: https://github.com/krzysiekfonal/grammaregex
# decoding of POS: https://www.guru99.com/pos-tagging-chunking-nltk.html

# adding additional named entity filters to pipe structure
# list of filters: https://stackoverflow.com/questions/59319207/ner-entity-recognition-country-filter

from textpipeliner import PipelineEngine, Context
from textpipeliner.pipes import *
import spacy

nlp = spacy.load('en_core_web_lg')
doc = nlp(content)

def VO_extractor(content):
    doc = nlp(content)
    pipes_structure = [
        FindTokensPipe("VERB"),
        AnyPipe([    
            SequencePipe([
                FindTokensPipe("VERB/dobj/*"),
                AggregatePipe([
                    NamedEntityFilterPipe("PERSON"),
                    NamedEntityFilterPipe("FAC"),
                    NamedEntityFilterPipe("LOC"),
                    NamedEntityFilterPipe("GPE"),
                    NamedEntityFilterPipe("PRODUCT"),
                    NamedEntityFilterPipe("LAW"),
                    NamedEntityFilterPipe("LANGUAGE"),
                    NamedEntityFilterPipe("DATE"),
                    NamedEntityFilterPipe("TIME"),
                    NamedEntityFilterPipe("PERCENT"),
                    NamedEntityFilterPipe("MONEY"),
                    NamedEntityFilterPipe("QUANTITY"),
                    NamedEntityFilterPipe("ORDINAL"),
                    NamedEntityFilterPipe("CARDINAL"),
                    NamedEntityFilterPipe("ORG"),
                    NamedEntityFilterPipe("WORK_OF_ART"),
                    NamedEntityFilterPipe("EVENT")
                    ]),
                NamedEntityExtractorPipe()]),
            SequencePipe([
                FindTokensPipe("VERB/nsubjpass/*"),
                AggregatePipe([
                    NamedEntityFilterPipe("PERSON"),
                    NamedEntityFilterPipe("FAC"),
                    NamedEntityFilterPipe("LOC"),
                    NamedEntityFilterPipe("GPE"),
                    NamedEntityFilterPipe("PRODUCT"),
                    NamedEntityFilterPipe("LAW"),
                    NamedEntityFilterPipe("LANGUAGE"),
                    NamedEntityFilterPipe("DATE"),
                    NamedEntityFilterPipe("TIME"),
                    NamedEntityFilterPipe("PERCENT"),
                    NamedEntityFilterPipe("MONEY"),
                    NamedEntityFilterPipe("QUANTITY"),
                    NamedEntityFilterPipe("ORDINAL"),
                    NamedEntityFilterPipe("CARDINAL"),
                    NamedEntityFilterPipe("ORG"),
                    NamedEntityFilterPipe("WORK_OF_ART"),
                    NamedEntityFilterPipe("EVENT")
                    ]),
                NamedEntityExtractorPipe()]),
            SequencePipe([
                FindTokensPipe("VERB/dobj/*")]),
            SequencePipe([
                FindTokensPipe("VERB/nsubjpass/*")])])
            ]
    engine = PipelineEngine(pipes_structure, Context(doc), [0, 1])
    VOs = engine.process()
    return VOs

# applying the VO_extractor function on each row of my pandas dataframe

term_list_df["VOs"] = term_list_df.apply(lambda row: VO_extractor(row["text"]), axis = 1)

VO_df = term_list_df[["ID", "Date", "News_source_name", "VOs"]] 

# within the dataframe, separating the list of VO pairs into individual VO pairs per row

VO_df_exploded = VO_df.explode("VOs")

VO_df_exploded[["V", "O"]] = pd.DataFrame(VO_df_exploded.VOs.tolist(), index = VO_df_exploded.index)

VO_df_exploded

# saving df with VO triples into a csv file

VO_df_exploded.to_csv("data/NOW_preprocessed/total_VO_df.csv", index = False)
