==> STEP 3. creating full dataset with SV, VO pairs, connotation frames, and political alignment ratings.ipynb <==
import pandas as pd


SV_df = pd.read_csv("data/NOW_preprocessed/total_SV_df.csv")
VO_df = pd.read_csv("data/NOW_preprocessed/total_VO_df.csv")

print(len(SV_df))
print(len(VO_df))

## Bringing SV, VO columns to a preprocessed format ##

# checking how many NAs there are

print(SV_df['SVs'].isna().sum()) 
print(VO_df['VOs'].isna().sum())

# removing the NAs

SV_df_nona = SV_df[SV_df['SVs'].notna()]
VO_df_nona = VO_df[VO_df['VOs'].notna()]


print(SV_df_nona['SVs'].isna().sum()) 
print(VO_df_nona['VOs'].isna().sum())

print(len(SV_df_nona))
print(len(VO_df_nona))

# taking the S, V, O columns out of list format, making it into string:

import re

def text_cleaner(text):
    text_cleaned = re.sub(r",", "", text) 
    return text_cleaned


# making the list column into string on the SV df

SV_df_nona['S_string'] = SV_df_nona['S'].apply(lambda x: x[1:-1])
SV_df_nona['S_string'] = SV_df_nona.apply(lambda row: text_cleaner(row["S_string"]), axis = 1)
SV_df_nona['V_string'] = SV_df_nona['V'].apply(lambda x: x[1:-1])

SV_df_nona


# making the list column into string on the VO df

VO_df_nona['V_string'] = VO_df_nona['V'].apply(lambda x: x[1:-1])
VO_df_nona['O_string'] = VO_df_nona['O'].apply(lambda x: x[1:-1])
VO_df_nona['O_string'] = VO_df_nona.apply(lambda row: text_cleaner(row["O_string"]), axis = 1)

VO_df_nona


### filtering for S and O that includes BLM relevant terms. Worthy revisions: extend term list, save to csv. #### Important to note: if term list fitlers are changed, the news sources have to be reassessed. Additional news sources might be inlcuded in the exported csv files, which have to be checked for bias ratings

# filtering to see how many BLM relevant Subjects there are in SV df
# adding a numeric code, representing the following: -1 = BLM side (against authority - liberal), 1 = anti-BLM side (for authority - conservative), 0 = shoudl trigger disapproval by both sides, e.g. Zimmerman, brutality, Chauvin

SV_df_nona_breonna_S = SV_df_nona[SV_df_nona['S_string'].str.contains("Breonna", case = False)]
SV_df_nona_breonna_S["Side"] = "-1"
SV_df_nona_floyd_S = SV_df_nona[SV_df_nona['S_string'].str.contains("Floyd", case = False)]
SV_df_nona_floyd_S["Side"] = "-1"
SV_df_nona_trayvon_S = SV_df_nona[SV_df_nona['S_string'].str.contains("Trayvon", case = False)]
SV_df_nona_trayvon_S["Side"] = "-1"
SV_df_nona_movement_S = SV_df_nona[SV_df_nona['S_string'].str.contains("movement", case = False)]
SV_df_nona_movement_S["Side"] = "-1"
SV_df_nona_blacklivesmatter_S = SV_df_nona[SV_df_nona['S_string'].str.contains("Black Lives Matter", case = False)]
SV_df_nona_blacklivesmatter_S["Side"] = "-1"
SV_df_nona_blm_S = SV_df_nona[SV_df_nona['S_string'].str.contains("BLM", case = False)]
SV_df_nona_blm_S["Side"] = "-1"
SV_df_nona_htblacklivesmatter_S = SV_df_nona[SV_df_nona['S_string'].str.contains("#BlackLivesMatter", case = False)]
SV_df_nona_htblacklivesmatter_S["Side"] = "-1"
SV_df_nona_protester_S = SV_df_nona[SV_df_nona['S_string'].str.contains("protester", case = False)]
SV_df_nona_protester_S["Side"] = "-1"
SV_df_nona_demonstrator_S = SV_df_nona[SV_df_nona['S_string'].str.contains("demonstrator", case = False)]
SV_df_nona_demonstrator_S["Side"] = "-1"
SV_df_nona_nationalguard_S = SV_df_nona[SV_df_nona['S_string'].str.contains("National Guard", case = False)]
SV_df_nona_nationalguard_S["Side"] = "1"
SV_df_nona_bluelivesmatter_S = SV_df_nona[SV_df_nona['S_string'].str.contains("Blue Lives Matter", case = False)]
SV_df_nona_bluelivesmatter_S["Side"] = "1"
SV_df_nona_htbluelivesmatter_S = SV_df_nona[SV_df_nona['S_string'].str.contains("#BlueLivesMatter", case = False)]
SV_df_nona_htbluelivesmatter_S["Side"] = "1"
SV_df_nona_alllivesmatter_S = SV_df_nona[SV_df_nona['S_string'].str.contains("All Lives Matter", case = False)]
SV_df_nona_alllivesmatter_S["Side"] = "1"
SV_df_nona_htalllivesmatter_S = SV_df_nona[SV_df_nona['S_string'].str.contains("#AllLivesMatter", case = False)]
SV_df_nona_htalllivesmatter_S["Side"] = "1"
SV_df_nona_police_S = SV_df_nona[SV_df_nona['S_string'].str.contains("police", case = False)]
SV_df_nona_police_S["Side"] = "1"
SV_df_nona_government_S = SV_df_nona[SV_df_nona['S_string'].str.contains("government", case = False)]
SV_df_nona_government_S["Side"] = "1"


SV_df_nona_terms = pd.concat([SV_df_nona_breonna_S,
                               SV_df_nona_floyd_S,
                               SV_df_nona_trayvon_S,
                               SV_df_nona_movement_S,
                               SV_df_nona_blacklivesmatter_S,
                               SV_df_nona_blm_S,
                               SV_df_nona_htblacklivesmatter_S,
                               SV_df_nona_protester_S,
                               SV_df_nona_demonstrator_S,
                               SV_df_nona_nationalguard_S,
                               SV_df_nona_bluelivesmatter_S,
                               SV_df_nona_htbluelivesmatter_S,
                               SV_df_nona_alllivesmatter_S,
                               SV_df_nona_htalllivesmatter_S,
                               SV_df_nona_police_S,
                               SV_df_nona_government_S
                              ],
                                 ignore_index=True)

SV_df_nona_terms

# filtering to see how many BLM relevant Objects there are in VO df
# adding a numeric code, representing the following: -1 = BLM side (against authority - liberal), 1 = anti-BLM side (for authority - conservative), 0 = shoudl trigger disapproval by both sides, e.g. Zimmerman, brutality, Chauvin

VO_df_nona_breonna_O = VO_df_nona[VO_df_nona['O_string'].str.contains("Breonna", case = False)]
VO_df_nona_breonna_O["Side"] = "-1"
VO_df_nona_floyd_O = VO_df_nona[VO_df_nona['O_string'].str.contains("Floyd", case = False)]
VO_df_nona_floyd_O["Side"] = "-1"
VO_df_nona_trayvon_O = VO_df_nona[VO_df_nona['O_string'].str.contains("Trayvon", case = False)]
VO_df_nona_trayvon_O["Side"] = "-1"
VO_df_nona_movement_O = VO_df_nona[VO_df_nona['O_string'].str.contains("movement", case = False)]
VO_df_nona_movement_O["Side"] = "-1"
VO_df_nona_blacklivesmatter_O = VO_df_nona[VO_df_nona['O_string'].str.contains("Black Lives Matter", case = False)]
VO_df_nona_blacklivesmatter_O["Side"] = "-1"
VO_df_nona_blm_O = VO_df_nona[VO_df_nona['O_string'].str.contains("BLM", case = False)]
VO_df_nona_blm_O["Side"] = "-1"
VO_df_nona_htblacklivesmatter_O = VO_df_nona[VO_df_nona['O_string'].str.contains("#BlackLivesMatter", case = False)]
VO_df_nona_htblacklivesmatter_O["Side"] = "-1"
VO_df_nona_protester_O = VO_df_nona[VO_df_nona['O_string'].str.contains("protester", case = False)]
VO_df_nona_protester_O["Side"] = "-1"
VO_df_nona_demonstrator_O = VO_df_nona[VO_df_nona['O_string'].str.contains("demonstrator", case = False)]
VO_df_nona_demonstrator_O["Side"] = "-1"
VO_df_nona_nationalguard_O = VO_df_nona[VO_df_nona['O_string'].str.contains("National Guard", case = False)]
VO_df_nona_nationalguard_O["Side"] = "1"
VO_df_nona_bluelivesmatter_O = VO_df_nona[VO_df_nona['O_string'].str.contains("Blue Lives Matter", case = False)]
VO_df_nona_bluelivesmatter_O["Side"] = "1"
VO_df_nona_htbluelivesmatter_O = VO_df_nona[VO_df_nona['O_string'].str.contains("#BlueLivesMatter", case = False)]
VO_df_nona_htbluelivesmatter_O["Side"] = "1"
VO_df_nona_alllivesmatter_O = VO_df_nona[VO_df_nona['O_string'].str.contains("All Lives Matter", case = False)]
VO_df_nona_alllivesmatter_O["Side"] = "1"
VO_df_nona_htalllivesmatter_O = VO_df_nona[VO_df_nona['O_string'].str.contains("#AllLivesMatter", case = False)]
VO_df_nona_htalllivesmatter_O["Side"] = "1"
VO_df_nona_police_O = VO_df_nona[VO_df_nona['O_string'].str.contains("police", case = False)]
VO_df_nona_police_O["Side"] = "1"
VO_df_nona_government_O = VO_df_nona[VO_df_nona['O_string'].str.contains("government", case = False)]
VO_df_nona_government_O["Side"] = "1"

VO_df_nona_terms = pd.concat([VO_df_nona_breonna_O,
                               VO_df_nona_floyd_O,
                               VO_df_nona_trayvon_O,
                               VO_df_nona_movement_O,
                               VO_df_nona_blacklivesmatter_O,
                               VO_df_nona_blm_O,
                               VO_df_nona_htblacklivesmatter_O,
                               VO_df_nona_protester_O,
                               VO_df_nona_demonstrator_O,
                               VO_df_nona_nationalguard_O,
                               VO_df_nona_bluelivesmatter_O,
                               VO_df_nona_htbluelivesmatter_O,
                               VO_df_nona_alllivesmatter_O,
                               VO_df_nona_htalllivesmatter_O,
                               VO_df_nona_police_O,
                               VO_df_nona_government_O
                              ],
                                 ignore_index=True)

VO_df_nona_terms

## Lemmatizing the V columns to enable the merger with connotation score dataset, and to allow for later analysis ##

# running on filtered term list dfs from above
# following this guide: https://www.geeksforgeeks.org/python-lemmatization-with-nltk/
# import these modules

from nltk.stem import WordNetLemmatizer
 
lemmatizer = WordNetLemmatizer()

def lemmatizer_fun(verb):
    verb_lemmatized = lemmatizer.lemmatize(verb, 'v')
    return verb_lemmatized

SV_df_nona_terms['V_lemmatized'] = SV_df_nona_terms.apply(lambda row: lemmatizer_fun(row["V_string"]), axis = 1)
VO_df_nona_terms['V_lemmatized'] = VO_df_nona_terms.apply(lambda row: lemmatizer_fun(row["V_string"]), axis = 1)

## Connotation dataset ##

# connotation scores originating from this paper: https://aclanthology.org/P16-1030.pdf

connotation_df = pd.read_csv("data/connotation_dataset/full_frame_info.txt", sep='\t')
connotation_df

# lemmatizing connotation score df to make sure it can be matched to the SV, VO dataframes by verb

# following this guide: https://www.geeksforgeeks.org/python-lemmatization-with-nltk/
# import these modules

from nltk.stem import WordNetLemmatizer
 
lemmatizer = WordNetLemmatizer()

def lemmatizer_fun(verb):
    verb_lemmatized = lemmatizer.lemmatize(verb, 'v')
    return verb_lemmatized

connotation_df['V_lemmatized'] = connotation_df.apply(lambda row: lemmatizer_fun(row["verb"]), axis = 1)

connotation_df

## Merging filtered SV, VO dfs with connotation frame scores

# running on filtered term list dfs from above

SV_df_nona_terms_connotation = SV_df_nona_terms.merge(connotation_df, on = 'V_lemmatized', how = 'left')

VO_df_nona_terms_connotation = VO_df_nona_terms.merge(connotation_df, on = 'V_lemmatized', how = 'left')

# running on filtered term list dfs from above

# removing rows that have no connotation rating for the verb

print(SV_df_nona_terms_connotation.verb.isna().sum())
print(VO_df_nona_terms_connotation.verb.isna().sum())

term_list_df_SV_connotation_nonaverb = SV_df_nona_terms_connotation[SV_df_nona_terms_connotation['verb'].notna()]
term_list_df_VO_connotation_nonaverb = VO_df_nona_terms_connotation[VO_df_nona_terms_connotation['verb'].notna()]

print(term_list_df_SV_connotation_nonaverb.verb.isna().sum())
print(term_list_df_VO_connotation_nonaverb.verb.isna().sum())

## News media bias sources

# reading in media bias df

media_bias_df = pd.read_csv("data/news_political_alignment_ratings/news_sources_matching_bestversion.csv", sep = ';')

media_bias_df

# renaming SV, VO dfs news source column to enable left join merger

term_list_df_SV_connotation_nonaverb = term_list_df_SV_connotation_nonaverb.rename(columns={'News_source_name': 'name_in_NOW_data'})
term_list_df_VO_connotation_nonaverb = term_list_df_VO_connotation_nonaverb.rename(columns={'News_source_name': 'name_in_NOW_data'})

print(len(term_list_df_SV_connotation_nonaverb))
print(len(term_list_df_VO_connotation_nonaverb))

# left joining media bias df to the SV, VO dataframes

SV_connotation_bias_df = term_list_df_SV_connotation_nonaverb.merge(media_bias_df, on = 'name_in_NOW_data', how = 'left')
VO_connotation_bias_df = term_list_df_VO_connotation_nonaverb.merge(media_bias_df, on = 'name_in_NOW_data', how = 'left')

SV_connotation_bias_df.to_csv("data/NOW_preprocessed/SV_BLM_bias_connotation_combined.csv", index = False)
VO_connotation_bias_df.to_csv("data/NOW_preprocessed/VO_BLM_bias_connotation_combined.csv", index = False)
